+++ RED/ViewModels/ControlCenter/SensorViewModel.cs
@@ -85,7 +85,7 @@ namespace RED.ViewModels.ControlCenter
             _cc.DataRouter.Subscribe(this, _cc.MetadataManager.GetId("Ultrasonic"));
         }
 
-        public void ReceiveFromRouter(byte dataId, byte[] data)
+        public void ReceiveFromRouter(ushort dataId, byte[] data)
         {
             switch (_cc.MetadataManager.GetTelemetry(dataId).Name)
             {
